Add cross-product

+++ src/procedures.f90
@@ -3,8 +3,8 @@ module procedures
 contains
 
     FUNCTION cross(a, b)
-  INTEGER, DIMENSION(3) :: cross
-  INTEGER, DIMENSION(3), INTENT(IN) :: a, b
+  real*8, DIMENSION(3) :: cross
+  real*8, DIMENSION(3), INTENT(IN) :: a, b
 
   cross(1) = a(2) * b(3) - a(3) * b(2)
   cross(2) = a(3) * b(1) - a(1) * b(3)
